Continue reworking process_file
The ultimate goal here is to reduce the "temporal complexity" of the
code, i.e., make it more obvious where the different values are set and
how they get passed from method to method.

The "process_file" construct is a bit distracting. I think removing it
to expose the metadata state (in abp_file_metadata) and the particular
OSM/ABIR readers will make the file interaction more straightforward.