Separate OSM Reading From File Locating
Getting the metadata about file paths was separated from reading the
.abpimage file but intricately linked to reading the .abposm file.

Separating these out into concrete steps makes it easier to reason about
bugs and the overall control flow of the initial file loading.

It's also enabling us to pick-and-choose which pieces of the project we
want to run when, which will be essential for saving a "workspace" and
resuming work with saved state.