Rework OSM_Reader to Remove State
This commit removes the OSM Reader maintaining state. This pushes the
state out into the SirveApp.cpp file, where it can eventually make its
way into the global application state.

This commit, mostly cherry-picked from PR #4, was updated to also affect
the calibration and external nuc .cpp files.

The separation of code and data is a prerequisite to being able to store
state in a workspace independently of how it was obtained.