Extract Config Loading from SirveApp
The configuration file (currently config/config.json) only needs read in
once and the code for this should be centralized. This change
centralizes all config loading and storage into a single struct on app
creation.

Note this introduces a very slight regression - namely, it used to be
the case that loading the ABIR frames would re-check the config file
each time to get the file version. I believe this was unintentional and
now force a single load of the file version. If we expect to be able to
change this config value during the operation of the tool, that
capability should be introduced as needed.

Note also that I left the config reading in the calibration code, as it
wasn't in the original PR #5. I may revisit this in the future.