Rework ABIR Reading to Single Function
This reworks the abir reader to have a single public function with a
single returned data type. This clarifies down the line what is allowed
to be accessed, when, which will help simplify the storage of pieces of
this program in workspaces.

When initially written, this commit highlighted that the ir_data
structure in the ABIR_Reader class was _never used_ and therefore was
removed. Now, it is being used in multiple places, so it has been left
as-is. This unfortunately leaves a "side effect" from the ABIR reading
(and a public variable) that will required dedicated effort to remove.

Aside from the above side-effect, this commit marks the total
integration of changes from PR #7.