Reduce Coupling Between Engineering Data/Plots
The Engineering_Data construct and the Engineering_Plots construct are
tightly coupled. This is an attempt to start reducing the amount of
places they're either duplicating purpose or interacting with each other
too often.

This commit addresses a slight bug, in that the UI sliders and frame
markers/etc. were exhibiting off-by-one errors that are now cleaned up.

This commit had to additionally address code that was not in the
original (PR #8) commit, e.g. the NUC_external_file code.

Additionally, a signal from the playback_controller initialization was
fired, which ended up hitting a slot in the ir_video (Video) that would
attempt to update the video on app load - before any video was loaded -
which was causing a crash at startup. Short-circuiting the video
updating in case there aren't any frames placed in the video solved that
issue.

Because this commit was layered on top of release 1.3 but written for
the previous code version, it may have introduced off-by-one errors in
fringe components, e.g. annotations or calibration or something. We'll
need to test it.