Rename and Simplify Plot UI Toggles
I think ultimately these should just trigger an action on the
Engineering_Plots class itself, but that implies the class maintains
control over its own UI elements and such and I'm not yet prepared to
extract the UI components.

There should also be a method/signal/etc. that ties the menu item icon
visibility to the toggle, but I can't do it all...

This commit finishes the work done in PR #8.