Add Backup Script
Since we can't easily sync to another environment yet, this script will
copy all of the cpp, h, pro, and ui files along with the icons and
config information to a folder, where it can then be zipped or moved
wherever needed.

All files are converted to a .txt extension to allow for max
portability.

To undo, the file ending _pro.txt should become .pro, the files ending
_ui.txt should become .ui, the files ending _h.txt should become .h, and
the remaining files ending .txt should become .cpp again.

Note - this commit finishes the work done in PR #9.