Add Vector for Processing States
This is not currently interactable but is a stub for future behavior.

Moving away from the weird "property->bool" mapping that limits us to 1
video per processing state is essential to allow viewing and interacting
with the history of processing steps.

The future (saving the workspace) will run through this dropdown and
vector.