Control Processing State Changes Through QComboBox
Please note the UI at this state is a bit broken - some of the "old"
control features like the checkboxes still exist and control state
changes - and a lot of code and data is stored redundantly in the "old"
methods. Ideally, a clean delete of the old control features (e.g. the
entire `something` vector) should now be feasible and the entire control
can flow through the new features.

The new capability chains events together in a vector. The chain of past
processing operations allow users to switch between any desired states
at their leisure.

This chain can now trivially be converted into workspace saves/loads!

This commit is equivalent to the end of PR #12.