Fix Intermittent Bug
A tricky interaction between the "states_cleared" signal, triggering the
QComboBox clear method, which then triggerred the QComboBox's
currentIndexChanged method, would try to access the array at index -1,
which was leading to an intermittent bug.

Note that this commit was added out of order - the original bug was not
fixed until PR #15 while the current stream of commits is now starting
to integrate changes from PR #13.