Remove Unused Code, Remove 'something' Vector
Now that the video_container has been reworked, a lot of the previous
code is unused.

After this was initially committed in PR #13, new references to the
`something` vector were added in version 1.3. The previous commit
removed some of those references, but this commit had to add some
additional changes to accomodate that removal. As a byproduct of the
`video_container` now being owned by the video_display class, I had to
cludge together a fix by allowing direct access to the current
processing state.

I don't love the fact that the video_container "leaks" its behavior like
this now, but don't want to mess too much with history yet to fix it. I
will leave this to clean up later.