Add and Use processing_state Struct
The entirety of a "processing state" is now contained in a new struct.
It contains both the legacy video_details object (which contains, for
example, the pixel cube itself) as well as the methodology used to
obtain that state.

Again, the awkward interaction between video_display and video_container
was patched for simplicity and is left to resolve at a later time.