Eliminate BinaryFileReader Memory Leak
There were a few "low-hanging" areas where memory was clearly being
leaked. This change reduces the memory use when a ByteSwap operation is
called on the binary file reader, which includes most OSM reading
operations as well as certain portions of the ABIR reader.

Without this change, simply reading in different image files (mostly
through the osm_load process, without using the abir_reader to read any
frames at all) would reliably lose about 8-10 MB of memory. Here is an
example of the memory footprint of Sirve without this change:

86.6MB - 95.8MB - 103.2MB - 109.5MB - 119.8MB

With the change, the same sequence of file loads present the following:

66.1MB - 67.6MB - 71.6MB - 70.3MB - 72.6MB