Eliminate abir_reader Memory Leak
This is another "low-hanging" memory leak that vastly improves the
memory usage of SIRVE over time.

Without this change, anytime portions of an ABIR file were loaded (for
example, when the user selected certain frames from a video), a ton of
memory proportional to the length of the frames was lost. This was
leading to rapid memory loss. Here is an example of the memory footprint
of SIRVE after opening a file and loading a certain number of frames:

OPERATION: openfile - 100f - 100f - 200f - 100f - 100f

MEMORY: 45.4 - 172.3 - 231.9 - 410.1 - 411.4 - 471.7

With the changes, the memory stays nearly constant:

MEMORY: 47.8 - 118.3 - 121.2 - 179.6 - 121.5 - 122.9

There are some more (small, like 1MB or so) memory leaks sprinkled
through the program - evident when, for example, layering various
filters on loaded frames and then reloading and repeating - which will
need future investigation.