Reduce Vector Reference Copying, Simplify Code
This change builds on the previous commit by completely reworking all
the code that used the get_xaxis_value function to something slightly
simpler.

This change is geared towards making it more obvious where the x-axis
index (which determines the x-axis value, depending on the choice of
x_axis_units) is used.

This is aimed at fixing off-by-one errors between all the various
graphs, charts, downloaded data, and ground truth.