Fix Off-By-One Errors
This tiny change fixes the off-by-one errors between the right and left
graphs by mimicking the code found in establish_plot_limits (basically,
taking in 0 as the index and only adding one in the case of frames).

This correctly sets sub_plot_xmin and sub_plot_xmax when the limits are
established and then uses the same logic to get the current_marker's x
value.

This does indicate that there are two blocks of code doing the same
thing!!