Add i_fov_x and i_fov_y plots
Allows plotting the i_fov_x and i_fov_y values for the sensor on the
right.

In the future, it may be desirable to put these values on the left frame
as well, in which case we'll have to revisit the choice here, but this
is a good-enough presentation for now.

The y-axis defaults (0.00001 to 0.00003) may also need to be changed,
I'm just not sure what the expected range of values should be.