Rework Video_Display's EnhancedLabel
The label was created with a size policy when the app started, but when
later videos were loaded (e.g. a 2nd OSM was created), the label
creation did not apply the same size policies.

This commit makes the label creation setup identical between the first &
nth creation of this field.