Add methods to disable EnhancedLabel zooming
Unfortunately the "zooming" logic built into the EnhancedLabel assumes
that the display is, pixel for pixel, identical to the underlying
content. This means that when allowing stretching or resizing of the
video content, things break in weird ways. It is probably simple to fix
by tracking the display's current scale factor in each direction but is
a problem for later if it is important to fix.

This commit simply adds a method to disable the label and another method
to reenable it. It is enabled by default.