Add Popout Video Display Feature
This commit integrates the popout feature into the app. It disables
zooming when the display is active but otherwise will retain the same
features as the original display, to include responsiveness to button
presses on the main app, annotation editing, color and overlay changes,
etc.

Note that it may be desirable to have a maintain/restore "original
aspect ratio" button or option here somehow. We can revisit later.