Fix Histogram Plot Title Bug
Turns out the histogram plot is duplicated between `chart` and
`chart_full` - and both are identical and contain the same data (and
actually are connected to the same slots), etc. However one was saying
it was the absolute histogram (the other was saying it was relative).

This became immediately obvious when popped out.