Allow the Histogram to Popout
Adds code allowing the user to popout the absolute histogram.

Note this uncovers a ... "bug", sorta, namely that the absolute
histogram is _different_ from the absolute histogram displayed when the
relative histogram is also shown.

This means if you pop out the absolute histogram, it ends up .... kinda
duplicated between the popout and the main window if you're looking at
the split (absolute/relative) view on the right.

In reality, it's not a duplicate - it's a completely separate chart that
has identical behavior and is connected to the same slots. That said,
it's confusing and unnecessary - both views of the absolute plot can use
the same Clickable_QChartView, just with a different size/size policy if
needed.

I think cleaning up the whole right-side plot window (removing the
checkbox from the left side & putting it on the right side, for example,
as well as getting rid of the QStackedLayout) would make this
significantly simpler to work with as we look at making changes to the
histogram.