Add Example of QTest Testing
This is an example of tests with QTest. It's built via a qmake block
(see the test {} logic inside the .pro file. The test block inside the
.pro file can be summoned by running:

..\5.15.2\.....\qmake.exe "CONFIG += test"

This will build the makefiles for testing mode ... then running nmake
(and nmake install at least once) will output a `run_all_tests`
executable inside the ..\tests directory. Simply executing that file
will run all the unit tests.

I don't much love this approach because it's a lot of steps & complexity
to end up with an executable. I'm not sure if there's an easier way,
though and repeated attempts with different approaches haven't shown me
anything more promising.

The approach that I took most directly was following this:
https://alexhuszagh.github.io/2016/using-qttest-effectively/

Additional enlightening reading can be found here:
https://community.kde.org/Guidelines_and_HOWTOs/UnitTests