Rework Bad Pixel Workspace Integration
I don't think bad pixels being a "processing state" is actually the
right way to think about this, since it doesn't alter the pixel values
themselves and exists outside of the processing chain.

This converts it to a separate piece of the workspace. Saving the
indeces (instead of just the yes/no flag) saves the work to reprocess it
on every reload. This is part of a larger effort to persist more
settings to the workspace.

As the workspace save/load behavior is getting longer and a fair bit
more convoluted (not to mention completely untested), I may want to
revisit it in the near future.