Give video display control to VideoDisplay class
The VideoDisplay class, oddly enough, has never actually controlled the
presentation of the ... video ... display.

This is the first in a series of minor changes geared to making
VideoDisplay the authoritative source of the video data. This will
remove weird kludges - like the video display and SirveApp class having
to do so much signal/slot communication - related to displaying the
video and associated data.

As one example, I can now move the "what frame, what time from midnight,
and what zulu time" description messages from SirveApp to VideoDisplay.
This will mean that SirveApp no longer has to also respond to the
playback controller updating a frame. VideoDisplay will be able to
handle that description updating on its own.