Fix Timing Offset Off-By-One Bug
Fix a pretty major bug in the timing offset code that misaligned the
VideoDisplay class' idea of the frame time by giving it the wrong subset
of frame data.

By simply setting the timing offset (through the Settings menu at the
top of the application), the VideoDisplay received an incorrect subset
of the plotting frame data. This meant that every single frame was
reporting the wrong az, el, time, etc. for the data as well as
incorrectly marking the tracks and other critical pieces of information.