Copy Video Label Display to VideoDisplay
The VideoDisplay class previously was only a holding container for the
actual QImage video itself. It would control, for example, metadata and
timing information that got put onto the QImage ... but did not control
the labels underneath of the image, such as the frame # and zulu time
labels.

As some of that information could end up on the QImage - and there was
redundant work being done to calculate this in a number of places,
including SirveApp - there were a lot of places where things were
difficult to keep in sync.

Additionally, it was difficult to control new displays of data (such as
the upcoming pinpoint feature) if the new data displays had to get
shipped back to SirveApp for their presentation and control.

This commit does not yet remove the (now-redundant) code from SirveApp
but provides a good foundation to build new display features.