Force Refresh on Processing State Change
This commit forces the video display to do a full refresh when the video
container changes the current state. Since this is governed by the
combobox, tying it to the signal is sufficient.

There was a bug here where setting the combobox too early (before the
ABIR load had finished, for example, setting the display_data in
VideoDisplay) caused a crash. This was caused by the video display
attempting to access all the things it nededed to display before they
were placed.

Moving the video container interaction around to after the ABIR load
fixed it in the weird edge case of "first load", which does technically
"change" the combo box index and therefore forces a refresh.