Add "Pinpoint" Feature
Adds a "pinpoint" feature to the app. This provides a button that, when
activated, will let you click on the canvas and select pixels. The
selected pixels will be persisted (up to the latest 3 pixels). The raw
value for those pixels will be displayed in a new box below the video.

This is a "new" sort of requirement for SIRVE - it has the VideoDisplay
class emitting its first signal, it's making the need for a
"left_click_state" sort of enum more obvious (the is_zoom_active/etc.
triad is a bit much), it's starting to stretch the way we're using the
zoom information, and altogether starting to push the boundaries.

In other news, turns out zooming all the way in makes it pretty obvious
that the repeated QImage scaled/etc. calls are making the pixels ...
decidedly non-square, so I'm not sure what to do about that. I may end
up needing to get rid of the entire "relative_zoom" chain and just use
the absolute zoom data.