Completely Rework Bad Pixel Logic
I lost the plot a bit and couldn't split out meaningful commits, so this
commit bundles the whole change into one.

Essentially, this commit changes the bad pixel process to impact the
first processing_state itself. Instead of being a list of pixels that is
stored alongside the states, it is now an intrinsic property of the
first state. This is because, ultimately, if the user chooses to
replace bad pixels, we want to straight up _replace_ the bad pixels with
a local (kernel) average. The only reason to persist the list of ones we
fixed is to be able to tell the user later which pixels were changed.

As the pixels themselves will be altered, any later filters (background
subtraction, etc.) will be based on the new (fake) value for the
replaced pixels. This will have implications on any
mathematical/scientific processing of the data.

Things done in this change include moving the bad pixel UI controls to
the processing tab, having a "confirmation" dialog when bad pixels are
replaced via the UI, renaming bad_pixel_identification to bad_pixels
since it now also includes replacement, and the change from a bad pixel
"mask" (a per-pixel 0 or 1 depending on whether it was good or bad,
respectively) into a list of bad pixel indeces.