Add capability to convert pinpoints to bad pixels
A new button provides the ability for the list of pinpointed pixels to
be converted to bad pixels. This comes with its own confirmation dialogs
and logic.

This does highlight the _lack_ of ability to distinguish why a certain
pixel is marked, which may come up as a need in the future.