Add capability to mark pixel as "good"
The inverse of the previous commit, pinpointed pixels can now be
manually marked as "good" if they were previously marked as bad.

Note that this change required a bit of rework. When a pixel is marked
as bad, its value is completely overwritten in the raw (original) data.
This means that in order to "un-mark" a pixel, you actually have to go
get the correct raw data again. This means you have to read the raw
pixel's data back out of the original file.

Since there isn't a good way to read only what we need, this essentially
forces the app to reset and go all the way back to the beginning to load
the full set of frames again. This has a lot of effects including zoom
reset and other shenanigans, but should be sufficient for now. In
theory, this feature shouldn't be used very often.