Fix OSM Loading Crash
This crash was caused by OSM loads attempting to delete lbl_pinpoint ...
after lbl_pinpoint had accidentally been deleted. This should never have
been deleted, as it is only created once.

This may address some intermittent crashes that were coming from the
pinpoint feature as well, as if the lbl_pinpoint was deleted, an update
that attempted to write to it might crash.