Draw track centroid box correctly when zoomed
The logic for writing a box around the centroid pixel was only ever
correct at the base zoom level. Repeated zooms, particularly when the
pixels became large, fixated the box on the top-left corner of the
pixel, as the box never resized and/or moved with the zooms.

This change makes it work, kinda, mostly.

It gets weird for the same reason the other zoom stuff gets weird -
pixels can be non-square and such. The entirety of SIRVE zooming needs
reworked at some point.