Extract video display "zoom" code to simple class
The "zoom" logic (code handling zoom states, zooming in, zooming out,
the location of pixels in the video through zoom states, the location of
clicks in the video through zoom states, etc.) was intricately tied to
the VideoDisplay class.

As a QWidget, the VideoDisplay class can't really be re-created easily
for testing and the logic can't easily be isolated for debugging.

Separating the zoom logic and all of the relevant calculations into a
simple class will enable testing of this code as well as simplifying the
overall "God class" of VideoDisplay.