Coerce zoom area to match aspect ratio
This doesn't alleviate all the non-square pixel concerns per the lengthy
comments left at certain points in the code, but it cleans up the code a
bit and makes it more obvious what is happening. Future manipulation of
the QImage into something that gets displayed in a more robust way will
still need to happen.