Introduce TrackDetails struct
By giving each frame the bare minimum thigns it needs to display the
tracks, namely the ID of the track and a way to get its centroid, we
introduce the simplest possible way for the VideoDisplay to add, modify,
update, and delete tracks by a unique identifier.

This change is aimed at reducing a ton of complexity around how tracks
were historically stored and used in the app.