Introduce TrackInformation class
As tracks will become a first class concern, they really need extracted
to their own home that can manage the creation, deletion, modification,
etc. thereof.

This doesn't completely finish the track refactor because a lot of the
track details (including the irradiance/etc. that gets plotted on the
right) is still intricately tied to some other classes and locations.