Switch track plotting to use TrackInformation
The plots on the right are a bit tangled between the frame/timing
information, the track irradiance/az/el information, and other assorted
pieces of the application.

This commit makes the irradiance graphs use the TrackInformation data.
It doesn't yet apply to the azimuth or elevation plots.

In order to untangle the code, it's important to localize the track
information and separate it from the other concerns.

This code structure isn't where I want to be long-term, but it's the
only way I can incrementally reduce these dependencies without breaking
everything.