Add track file reading with error handling
Adds a rudimentary track CSV file reading capability. Nothing is yet
being done with the returned tracks.

Note that I chose to implement "read a track file" and "add tracks to
the application" as two separate steps. This is because the application
may want to do things depending on the track results in the file, for
example:
 - warn the user if a track ID conflicts with an existing manual track
 - give tracks a new ID if desired
 - select which tracks to actually import
 - etc.

By splitting out "read track file" and then a later "add tracks"
capability, the code is flexible to handle anything that may come my
way.