Persist manual tracks and check for replacement
Now, the manual tracks will be persisted inside of the TrackInformation
class for later use. The IDs that are ingested will also be tracked. The
SirveApp will check against the "known" manual IDs and warn the user
that it is overwriting a manual ID. Instead of a warning, it may be
desirable in the future to make this a confirmation dialog, selectable
list of "which ones do I want to import," or something else.