Display manual tracks on video display
Now that the manual tracks can be loaded, they can be displayed on the
primary video display.

Note that this change means VideoDisplay assumes it has the
manual_track_frames set and queryable, which is why they must be
provided (even if empty) on first ABIR load.

Future changes will presumably allow color customization as well as
selecting which manual tracks to show (per-color, even, we may get crazy
...)