Create logic outline for manual track creation
Prepares the main application for the manual track creation as well as
implements all the expected error, warning, and success cases. The main
logic left to implement is the actual video display controls and logic
to manipulate a manual track, then the save/overwrite logic.

The track management widget could also get an edit button for each track
that kicks off the same process.