Reduce external calls to update_display_frame
This doesn't get us all the way there, but in an ideal future, the
video_display object should have full control over knowing when it needs
to refresh. The main application should be able to modify it and trust
that the display is updated as needed.

This will allow a future where partial refreshes (i.e. only redrawing
the annotations and not the underlying frame, for example) are possible,
as the video_display will know what things need what level of refresh.
Forcing all that logic inside video_display is a good start.