Remove QStackedLayout, histogram chart duplication
The histogram plots were previously divided into two states in a
QStackedLayout, specifically:
1. A pane with the absolute histogram
2. A pane with a different absolute histogram and a relative histogram

By toggling between the layouts, you would actually be switching between
two different absolute histograms that just happened to be identical.
This was leading to confusing code and weird behavior, e.g. when you
popped out the absolute histogram from pane 1 and then it still appeared
docked alongside the relative histogram when toggled to pane 2.

This change removes the duplication. The relative histogram is now
added/removed from the single remaining layout as needed.

Note that we may need to look at the SizePolicy/sizeHint of the layout
and QFrames involved here to ensure a consistent display.