Consolidate and simplify HistogramLine_Plot code
Previously, there was a lot of duplicated code based around the idea
that you could avoid rendering the relative histogram if it wasn't
needed. This added complexity to the app by forcing a refresh on
VideoDisplay if the relative plot was toggled on, and wasn't saving much
time given that the code was nearly identical.

This commit greatly reduces the quantity and complexity of code in the
HistogramLine_Plot class.