Add unit tests verifying az/el calculation
The az/el calculation can be unit tested to confirm the correctness of
the math, but it highlights a particularly tricky issue right now -
specifically, the reported OSM track centroids in the data export and
the "visible" OSM track centroid boxes are not the same. These unit
tests don't make a ton of sense right now because of this discrepancy -
a manual track occupying the same pixels as the OSM track would have a
different az/el result.

The root of this issue hasn't yet been discovered but, as the math is
verifiably right, work will continue apace.