Add basic plotting of manual track az/el
This adds the minimal "plotting" functionality required for the manual
tracks. The manual track az/el numbers now appear on the right side
plots and correctly reflect additions, edits, or deletions to the manual
tracks in the app.

There is one major issue with this feature right now, namely the current
plotting capabilities (e.g. add_series) gives us no way to control the
color of these lines. Especially when it is "broken" (i.e. there's a gap
in the middle of the track), the series will have >1 color. As we look
to add per-track colors, this will have to be reworked.