Clean up PopoutDialog, fix intermittent crash
There was an intermittent crash caused by comparing the popout_video
pointer to nullptr in cases where it didn't exist (a non-existent object
in C++ doesn't guarantee that a pointer == nullptr).

This commit reworks the PopoutDialog to be a persistent object so that
we can use the built in isVisible() method to check if the underlying
widget is currently popped out.