Fixed Histogram, removed alpha
I think I have finally found a good solution for the histogram output until we really fix the scale issues. I also removed the alpha values in the RGB stylesheet because I was seeing warnings about it. I think some parameter has to be set in order for the alpha to work.