Fixed processing state integration
I think I fixed the processing state integration so that offsets are recorded in the json file and so that they can be maintained for future requirements.  We have to somehow pull them and adjust any interaction with the shifted frames to account for the offsets. I.e. pinpoint, osm tracks, etc.