Added GUI Reset function
Added a function to delete generated track info and clear the controls when a new frame range is loaded. The function also makes sure the video slider is set to 0 before trying to load frames as this was causing a crash when the slider was positioned outside of the new frame range.  (Might want to rename the function since it mostly only deals with the tracks but it could be a general function to reset the gui to an initial state.)  Also include some minor GUI updates and refactors.