Refactored and improved autotracking
Refactored autotracking with member functions. Changed to measure the actual frame values instead of the 8 bit values. Outputs peak counts, total counts above threshold, total counts in the rectangular ROI, the number of pixels above threshold, and the number of pixels in the ROI.  Also has logic to let the user select a new ROI if the track is lost (peak intensity is less than 50% of previous).