Updated track data export
Moved the track data export to the main window menu under export. There was an error in the plot option menu where we were redefining an action option to change the banner so it wasn't being displayed. I fixed this, too. I added an auto filename generator to the track data export. I also added the intensity counts to the output for manual and auto tracks. I also changed the background color of the plot space.