Fixes the autotracking lost track parameter and fixes the cmb_manual_track_IDs
I went the wrong direction on the autotrack lost parameter so it is corrected, here. I fixed the cmb_manual_track_IDs method to populate the items to be consistent with what is done elsewhere in the code. This also fixes a problem I had with the other method that didn't sort track numbers higher than 10 properly.