Fixes an oversight with the offsets upon track deletion
The offsets were not being deleted in the image when a manual track was deleted.  Therefore, centering on tracks would still see those old offsets. Old use case... import two tracks and center on both ("primary"). Then delete the first track and try to center on the primary, again (from uncentered state). You will see that the image still gets centered on the first track even though it is gone.  This fixes that.