Refactored auto track to remove extra file read
Refactored the auto track so that it no longer reads the track info back in from the .csv file. Instead, it generates the track details structure from the returned matrix. This is more consistent with the manual track creation, removes an unnecessary file read, and avoids the intermittent bug caused when the file is formatted with exponential notation.