Edge bug fix with improved opencv windows
I believe this fixes a bug that was causing crashes if you tried to track an object that leaves the frame boundary in frames centered on a *different* object.  Also includes updates opencv windows that are consistently positioned and includes an additional window that shows the track in the raw data frame.