bug fixes and significant changes to tracking
Modified the tracking to sue the filtered images.  Should have been that way to begin with. Also made corrections so that the tracking works and displays correctly with the D data.  Questionable change to remove negative values from sum counts calculations.  This seems to make the calcs work for D data and doesn't seem to affect the B Data calcs but need to verify.