Fixed video save bug
The video save feature was not working at all. I think I have that fixed and I changed the behavior a little so that playback stops when the recording stops and so that changing the fps doesn't start playback. I can't get the record/stop recording button to maintain its size.  Don't understand why.